+++ src/NewPAge/CrewPage.js
@@ -1,38 +1,5 @@
 import React, { useState, useEffect, useReducer } from "react";
-export const crews = [
-  {
-    id: "01",
-    position: "COMMANDER",
-    name: "DOUGLAS HURLEY",
-    about:
-      "Douglas Gerald Hurley is an American engineer, former Marine Corps pilot and former NASA astronaut. He launched into space for the third time as commander of Crew Dragon Demo-2.",
-    photo: "/assets/crew/image-douglas-hurley.png",
-  },
-  {
-    id: "02",
-    position: "MISSION SPECIALIST",
-    name: "MARK SHUTTLEWORTH",
-    about:
-      "Mark Richard Shuttleworth is the founder and CEO of Canonical, the company behind the Linux-based Ubuntu operating system. Shuttleworth became the first South African to travel to space as a space tourist.",
-    photo: "/assets/crew/image-mark-shuttleworth.webp",
-  },
-  {
-    id: "03",
-    position: "PILOT",
-    name: "VICTOR GLOVER",
-    about:
-      "Pilot on the first operational flight of the SpaceX Crew Dragon to the International Space Station. Glover is a commander in the U.S. Navy where he pilots an F/A-18.He was a crew member of Expedition 64, and served as a station systems flight engineer. ",
-    photo: "/assets/crew/image-victor-glover.webp",
-  },
-  {
-    id: "04",
-    position: "FlIGHT ENGINEER",
-    name: "ANOUSHEH ANSARI",
-    about:
-      "Anousheh Ansari is an Iranian American engineer and co-founder of Prodea Systems. Ansari was the fourth self-funded space tourist, the first self-funded woman to fly to the ISS, and the first Iranian in space. ",
-    photo: "/assets/crew/image-anousheh-ansari.webp",
-  },
-];
+import { crews } from "../components/Constant";
 
 export default function CrewPage() {
   const [selectedIndex, setSelectedIndex] = useState(0);
@@ -50,11 +17,7 @@ export default function CrewPage() {
 
   const scrollToCrew = (index) => {
     setSelectedIndex(index);
-    ref[index].current.scrollIntoView({
-      // behavior: "smooth",
-      // block: "nearest",
-      // inline: "start",
-    });
+    ref[index].current.scrollIntoView({});
   };
   return (
     // parent div
@@ -75,8 +38,8 @@ export default function CrewPage() {
       </div>
       {/* crew image carousel arnold*/}
       {/* flex  */}
-      <div id="mainPutang " className="lg:flex lg:ml-10">
-        <div id="putang1" className="lg:-mt-[150px] lg:ml-6">
+      <div className="lg:flex lg:ml-10">
+        <div className="lg:-mt-[150px] lg:ml-6">
           <div className="md:absolute md:mt-[300px]">
             <div className="flex overflow-x-hidden w-full ">
               {crews.map((crew, index) => (
@@ -120,10 +83,7 @@ export default function CrewPage() {
         </div>
 
         {/*  crew position */}
-        <div
-          id="putang2"
-          className="  lg:justify-center lg:hidden md:pt-[90px]"
-        >
+        <div className="  lg:justify-center lg:hidden md:pt-[90px]">
           <div className="md:to-blue-600  md:absolute md:mb-[900px] md:ml-[100px] md:w-[592px] lg:mt-[300px]">
             <div className=" flex justify-center items-center  font-bellefair text-[16px] pt-[50px] text-gray-200">
               <div className="md:font-bellefair md:text-[24px]">
